fix pkscript parsing
