--- blockchain/model/address.go
@@ -1,5 +1,0 @@
-package model
-
-type Address struct {
-	Encoded string
-}
