+++ go.mod
@@ -11,6 +11,8 @@ require (
 	github.com/lbryio/lbcutil v1.0.202-rc3
 	github.com/lbryio/lbry.go/v3 v3.0.0-20220114204055-99b18d4b5dfe
 	github.com/lbryio/types v0.0.0-20201019032447-f0b4476ef386
+	github.com/mmcloughlin/profile v0.1.1 // indirect
+	github.com/pkg/profile v1.6.0 // indirect
 	github.com/sirupsen/logrus v1.8.1
 	github.com/syndtr/goleveldb v1.0.1-0.20210819022825-2ae1ddf74ef7
 	golang.org/x/crypto v0.0.0-20211209193657-4570a0811e8b
